requirements.txt
discord.py
pydantic-settings
